backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.5
Flask-Limiter==3.5.0
Flask-SocketIO==5.3.5
Flask-Mail==0.9.1
Flask-WTF==1.2.1
python-socketio==5.10.0
python-engineio==4.8.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
APScheduler==3.10.4
Alembic==1.12.1
sentry-sdk==1.39.1
prometheus-client==0.19.0
prometheus-flask-exporter==0.23.0
psutil==5.9.6
redis==5.0.1
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
